+++ src/android/app/src/main/java/io/github/alexdenton/interviewd/api/repositories/InterviewRepository.kt
@@ -7,9 +7,9 @@ import io.reactivex.Single
  * Created by ryan on 9/9/17.
  */
 interface InterviewRepository {
-    fun getInterview(id: Int): Single<Interview>
+    fun getInterview(id: Long): Single<Interview>
     fun getAllInterviews(): Single<List<Interview>>
     fun createInterview(interview: Interview): Single<Interview>
     fun updateInterview(interview: Interview): Single<Interview>
-    fun deleteInterview(id: Int): Single<Interview>
+    fun deleteInterview(id: Long): Single<Interview>
 }
