+++ src/android/app/src/main/java/io/github/alexdenton/interviewd/api/repositories/TemplateRepository.kt
@@ -7,9 +7,9 @@ import io.reactivex.Single
  * Created by ryan on 8/14/17.
  */
 interface TemplateRepository {
-    fun getTemplate(id: Int): Single<Template>
+    fun getTemplate(id: Long): Single<Template>
     fun getAllTemplates(): Single<List<Template>>
     fun createTemplate(template: Template): Single<Template>
     fun updateTemplate(template: Template): Single<Template>
-    fun deleteTemplate(id: Int): Single<Template>
+    fun deleteTemplate(id: Long): Single<Template>
 }
